Initial structure.
Initialises a lerna repository with two web apps: a client, and an admin app. These are simple React applications, but they also utilise a shared common library. In this library are some react components.

Webpack tree shaking is enabled too, so that parts of the common library not used by a particular web app aren't built into their distributable